Revert file name changes and adds Auth0
