+++ server/frontend/package.json
@@ -10,17 +10,20 @@
     "react-scripts": "0.9.0"
   },
   "proxy": "http://localhost:5000",
-   "dependencies": {
-      "bootstrap": "^3.3.7",
-      "classnames": "^2.2.5",
-      "express": "^4.14.0",
-      "nectarine": "0.0.9",
-      "react": "^15.4.2",
-      "react-addons-pure-render-mixin": "^15.4.2",
-      "react-bootstrap": "^0.30.8",
-      "react-dom": "^15.4.2",
-      "react-router": "^3.0.2",
-      "socket.io": "^1.7.2",
+  "dependencies": {
+    "auth0-lock": "^10.14.0",
+    "bootstrap": "^3.3.7",
+    "classnames": "^2.2.5",
+    "express": "^4.14.0",
+    "jwt-decode": "^2.2.0",
+    "nectarine": "0.0.9",
+    "react": "^15.4.2",
+    "react-addons-pure-render-mixin": "^15.4.2",
+    "react-bootstrap": "^0.30.8",
+    "react-dom": "^15.4.2",
+    "react-router": "^2.8.0",
+    "react-router-bootstrap": "^0.23.1",
+    "socket.io": "^1.7.2"
   },
   "scripts": {
     "start": "react-scripts start",
