+++ server/frontend/public/index.html
@@ -13,7 +13,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>SPOT</title>
+    <title>SPOTS</title>
   </head>
   <body>
     <div id="root"></div>
