+++ server/frontend/src/containers/Error404/index.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class Error404 extends Component {
+  render () {
+    return (
+      <div className="Error-Class">
+        <h1>Error 404: You shall not pass!</h1>
+      </div>
+    )
+  }
+}
+
+export default Error404;
