+++ server/frontend/src/containers/Home/index.js
@@ -1,14 +1,13 @@
 //src/containers/Login/index.js
 import React, { Component } from 'react';
-import classnames from 'classnames';
 
 class Home extends Component {
 
   render() {
     return (
       <div>
-	  	<h1> Hello from Home! </h1>
-	  </div>
+        <h1> Hello from Home! </h1>
+      </div>
     );
   }
 }
