+++ server/frontend/src/containers/Register/index.js
@@ -4,12 +4,12 @@ import React, { Component } from 'react';
 import classnames from 'classnames';
 
 class Register extends Component {
+
   render () {
     const { className, ...props } = this.props;
-
     return (
-      <div className={classnames('Register', className)} {...props}>
-        <h2>Hello from Register!</h2>
+      <div className="Register">
+        <h1> Hello from REgister </h1>
       </div>
     );
   }
