+++ server/frontend/src/containers/ViewAccount/index.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class ViewAccount extends Component {
+	render () {
+		return (
+			<div className="ViewAccount-Class">
+				<h1>Hello from viewing accounts</h1>
+			</div>
+		);
+	}
+}
+
+export default ViewAccount;
